+++ src/components/shared/badges/MeetingDate.tsx
@@ -14,7 +14,9 @@ function formatDate(dateString: string) {
 
 const MeetingDate: React.FC<MeetingDateProps> = ({ dateTime }) => {
   const formatted = formatDate(dateTime);
-  return <span className="text-gray-800 text-sm">{formatted}</span>;
+  return (
+    <span className="rounded bg-white text-sm text-gray-800">{formatted}</span>
+  );
 };
 
 export default MeetingDate;
